requirements.txt
torchsummary==1.5.1
torch==1.3.1
torchvision==0.4.2
numpy==1.16.2
opencv_python==3.4.3.18
Pillow==6.2.1
